requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.3.3
langchain-google-genai==2.0.0
google-generativeai==0.7.2
tenacity==8.5.0
langchain-groq>=0.1.0
